Reverting to an auto-incrementing patch version
Tasks can only have up to 3 segments in their version number and Pipelines seems to keep a cache of tasks based on version number. So without updating the patch version, Azure will continue to use the cached version irrespective of whether the task has actually changed. So we have no choice but to increment the patch.